back/requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==3.0.2
Pygments==2.5.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
